pages/Pagina_principal/requirements.txt
streamlit
geopandas
scipy
matplotlib
seaborn
plotly